# web/openapi.json
{
  "components": {
    "schemas": {
...
      "CommonResponse": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse",
        "type": "object"
      },
...
      "CommonResponse_Any_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[Any]",
        "type": "object"
      },
...
      "CommonResponse_CreateWorkProjectSessionResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CreateWorkProjectSessionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[CreateWorkProjectSessionResponse]",
        "type": "object"
      },
...
      "CommonResponse_DeleteWorkProjectResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeleteWorkProjectResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[DeleteWorkProjectResponse]",
        "type": "object"
      },
...
      "CommonResponse_ListWorkProjectSessionsResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ListWorkProjectSessionsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[ListWorkProjectSessionsResponse]",
        "type": "object"
      },
...
      "CommonResponse_QueryWorkProjectsResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueryWorkProjectsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[QueryWorkProjectsResponse]",
        "type": "object"
      },
...
      "CommonResponse_WorkProjectRecordSnapshotSchema_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkProjectRecordSnapshotSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[WorkProjectRecordSnapshotSchema]",
        "type": "object"
      },
      "CommonResponse_WorkProjectSchema_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkProjectSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[WorkProjectSchema]",
        "type": "object"
      },
...
      "CreateWorkProjectRequest": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/WorkProjectAssetRequest"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Assets",
            "type": "array"
          },
          "description": {
            "default": "",
            "maxLength": 2000,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "owner_user_ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 100,
            "title": "Owner User Ids",
            "type": "array"
          },
          "sandbox_container_ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 100,
            "title": "Sandbox Container Ids",
            "type": "array"
          },
          "type": {
            "$ref": "#/components/schemas/WorkProjectType",
            "default": "penetration_test"
          }
        },
        "required": [
          "name",
          "assets"
        ],
        "title": "CreateWorkProjectRequest",
        "type": "object"
      },
...
      "DeleteWorkProjectResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id"
        ],
        "title": "DeleteWorkProjectResponse",
        "type": "object"
      },
...
      "QueryWorkProjectsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkProjectSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "size",
          "total",
          "items"
        ],
        "title": "QueryWorkProjectsResponse",
        "type": "object"
      },
...
      "UpdateWorkProjectMetadataRequest": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/WorkProjectAssetRequest"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Assets",
            "type": "array"
          },
          "description": {
            "default": "",
            "maxLength": 2000,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "owner_user_ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 100,
            "title": "Owner User Ids",
            "type": "array"
          },
          "sandbox_container_ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 100,
            "title": "Sandbox Container Ids",
            "type": "array"
          },
          "type": {
            "$ref": "#/components/schemas/WorkProjectType",
            "default": "penetration_test"
          }
        },
        "required": [
          "name",
          "assets"
        ],
        "title": "UpdateWorkProjectMetadataRequest",
        "type": "object"
      },
...
      "WorkProjectType": {
        "enum": [
          "penetration_test",
          "source_code_audit"
        ],
        "title": "WorkProjectType",
        "type": "string"
      }
    },
    "securitySchemes": {
      "AccessTokenAuth": {
        "in": "header",
        "name": "X-Z3r0-Access-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Z3r0 - A controlled multi-agent workbench for authorized security assessment, code auditing, internal review, and controlled research.",
    "version": "0.2.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/work-projects": {
      "get": {
        "operationId": "query_work_projects_route_api_work_projects_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "title": "Keyword",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_QueryWorkProjectsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Query Work Projects Route",
        "tags": [
          "work-projects"
        ]
      },
      "post": {
        "operationId": "create_work_project_route_api_work_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_WorkProjectSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Create Work Project Route",
        "tags": [
          "work-projects"
        ]
      }
    },
    "/api/work-projects/{id}": {
      "delete": {
        "operationId": "delete_work_project_handler_api_work_projects__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_DeleteWorkProjectResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Work project not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Delete Work Project Handler",
        "tags": [
          "work-projects"
        ]
      }
    },
    "/api/work-projects/{id}/cancel": {
      "post": {
        "operationId": "cancel_work_project_handler_api_work_projects__id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_WorkProjectSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Work project not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Cancel Work Project Handler",
        "tags": [
          "work-projects"
        ]
      }
    },
    "/api/work-projects/{id}/metadata": {
      "patch": {
        "operationId": "update_work_project_metadata_route_api_work_projects__id__metadata_patch",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkProjectMetadataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_WorkProjectSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Work project not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Update Work Project Metadata Route",
        "tags": [
          "work-projects"
        ]
      }
    },
    "/api/work-projects/{id}/record-snapshot": {
      "get": {
        "operationId": "get_work_project_record_snapshot_route_api_work_projects__id__record_snapshot_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_WorkProjectRecordSnapshotSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Work project not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Get Work Project Record Snapshot Route",
        "tags": [
          "work-projects"
        ]
      }
    },
    "/api/work-projects/{id}/retry": {
      "post": {
        "operationId": "retry_work_project_handler_api_work_projects__id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_WorkProjectSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Work project not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Retry Work Project Handler",
        "tags": [
          "work-projects"
        ]
      }
    },
    "/api/work-projects/{id}/sessions": {
      "get": {
        "operationId": "list_work_project_sessions_route_api_work_projects__id__sessions_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_ListWorkProjectSessionsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Work project not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "List Work Project Sessions Route",
        "tags": [
          "work-projects"
        ]
      },
      "post": {
        "operationId": "create_work_project_session_route_api_work_projects__id__sessions_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_CreateWorkProjectSessionResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Work project not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Create Work Project Session Route",
        "tags": [
          "work-projects"
        ]
      }
    },
    "/api/work-projects/{id}/sessions/{session_id}": {
      "delete": {
        "operationId": "delete_work_project_session_route_api_work_projects__id__sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Work project not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Delete Work Project Session Route",
        "tags": [
          "work-projects"
        ]
      }
    }
  }
}